smoking/requirements.txt
ultralytics==8.2.77
flask
opencv-python
pillow
